chore: project restructuring and fixes
