+++ src/components/Tools/index.ts
@@ -3,3 +3,6 @@ export { default as Chmod } from './Chmod';
 
 // export * from './ToDo';
 // export { default as ToDo } from './ToDo';
+
+export * from './Workspace';
+export { default as Workspace } from './Workspace';
